refactor/src/change-time-params-from-int-to-float (#9)

+++ src/visiongui/driver/exception.py
@@ -2,7 +2,7 @@ class ExceptionTimeout(Exception):
     def __init__(
         self,
         *,
-        timeout: int,
+        timeout: float,
     ):
         super().__init__(f"Element was not found within the {timeout}-second timeout")
 
@@ -11,8 +11,8 @@ class ExceptionElementNotStableLongEnough(Exception):
     def __init__(
         self,
         *,
-        time_held_stable_on_screen: int,
-        timeout: int,
+        time_held_stable_on_screen: float,
+        timeout: float,
     ):
         super().__init__(
             f"Element was not stable on screen for {time_held_stable_on_screen} seconds within the {timeout}-second timeout",
@@ -23,7 +23,7 @@ class ExceptionElementNotFound(ExceptionTimeout):
     def __init__(
         self,
         *,
-        timeout: int,
+        timeout: float,
     ):
         super().__init__(timeout=timeout)
 
@@ -33,7 +33,7 @@ class ExceptionWindowNotFound(Exception):
         self,
         *,
         window_title: str,
-        timeout: int,
+        timeout: float,
     ):
         super().__init__(
             f"Window '{window_title}' not found within the {timeout}-second timeout",
